refact: move dynamic_import and mako_require to visitors
