+++ crates/mako/src/compiler.rs
@@ -287,15 +287,12 @@ impl Compiler {
         }
 
         if !config.ignores.is_empty() {
-            let ignore_regex = config
+            let ignores = config
                 .ignores
                 .iter()
                 .map(|ignore| Regex::new(ignore).map_err(Error::new))
                 .collect::<Result<Vec<Regex>>>()?;
-
-            plugins.push(Arc::new(plugins::ignore::IgnorePlugin {
-                ignores: ignore_regex,
-            }))
+            plugins.push(Arc::new(plugins::ignore::IgnorePlugin { ignores }))
         }
 
         let plugin_driver = PluginDriver::new(plugins);
