+++ crates/mako/src/transform_in_generate.rs
@@ -3,7 +3,8 @@ use std::sync::mpsc::channel;
 use std::sync::Arc;
 use std::time::Instant;
 
-use mako_core::anyhow::Result;
+use mako_core::anyhow::{Error, Result};
+use mako_core::regex::Regex;
 use mako_core::swc_common::errors::HANDLER;
 use mako_core::swc_common::GLOBALS;
 use mako_core::swc_css_ast;
@@ -22,16 +23,15 @@ use mako_core::tracing::debug;
 
 use crate::ast_2::js_ast::JsAst;
 use crate::compiler::{Compiler, Context};
-use crate::config::OutputMode;
 use crate::module::{Dependency, ModuleAst, ModuleId, ResolveType};
 use crate::thread_pool;
 use crate::transformers::transform_async_module::{mark_async, AsyncModule};
-use crate::transformers::transform_dynamic_import::DynamicImport;
-use crate::transformers::transform_mako_require::MakoRequire;
 use crate::transformers::transform_meta_url_replacer::MetaUrlReplacer;
 use crate::transformers::transform_optimize_define_utils::OptimizeDefineUtils;
 use crate::visitors::css_imports::CSSImports;
 use crate::visitors::dep_replacer::{DepReplacer, DependenciesToReplace};
+use crate::visitors::dynamic_import::DynamicImport;
+use crate::visitors::mako_require::MakoRequire;
 
 impl Compiler {
     pub fn transform_all(&self, async_deps_map: HashMap<ModuleId, Vec<Dependency>>) -> Result<()> {
@@ -240,14 +240,23 @@ pub fn transform_js_generate(transform_js_param: TransformJsParam) -> Result<()>
                         let mut meta_url_replacer = MetaUrlReplacer {};
                         ast.ast.visit_mut_with(&mut meta_url_replacer);
 
-                        let mut dynamic_import = DynamicImport { context };
+                        let mut dynamic_import = DynamicImport {
+                            context: context.clone(),
+                        };
                         ast.ast.visit_mut_with(&mut dynamic_import);
 
-                        // replace require to __mako_require__ for bundle mode
-                        if matches!(context.config.output.mode, OutputMode::Bundle) {
-                            let mut mako_require = MakoRequire::new(context, unresolved_mark);
-                            ast.ast.visit_mut_with(&mut mako_require);
-                        }
+                        // replace require to __mako_require__
+                        let ignores = context
+                            .config
+                            .ignores
+                            .iter()
+                            .map(|ignore| Regex::new(ignore).map_err(Error::new))
+                            .collect::<Result<Vec<Regex>>>()?;
+                        let mut mako_require = MakoRequire {
+                            ignores,
+                            unresolved_mark,
+                        };
+                        ast.ast.visit_mut_with(&mut mako_require);
 
                         ast.ast
                             .visit_mut_with(&mut hygiene_with_config(hygiene::Config {
