+++ crates/mako/src/transformers/mod.rs
@@ -1,8 +1,5 @@
 pub mod transform_async_module;
-pub mod transform_dynamic_import;
-pub mod transform_mako_require;
 pub mod transform_meta_url_replacer;
 pub mod transform_optimize_define_utils;
 // pub mod transform_optimize_package_imports;
 pub mod transform_px2rem;
-pub mod utils;
