+++ crates/mako/src/visitors/mod.rs
@@ -5,8 +5,10 @@ pub(crate) mod css_imports;
 pub(crate) mod default_export_namer;
 pub(crate) mod dep_analyzer;
 pub(crate) mod dep_replacer;
+pub(crate) mod dynamic_import;
 pub(crate) mod dynamic_import_to_require;
 pub(crate) mod env_replacer;
+pub(crate) mod mako_require;
 pub(crate) mod provide;
 pub(crate) mod react;
 pub(crate) mod try_resolve;
